Removed addition of GO_PIPELINE_LABEL to version.
Until now a version name has been put into the version branch that does not have the GoCD build counter in it. Later on the file is read and the counter appended to it before it is used in Docker tags.
This does not allow for version names that do not include the GoCD build counter. With release candidates the counter is not necessary.
Changing the file to include the complete version name (including the GoCD build counter if necessary) means that it can be used for other applications.