Several enhancements to the Makefile.
Version can be passed at the command line to override the reading of `version_branch.txt`.
Running PIP from the venv works on both Windows and Linux.
The `get_sql_dump` target is now a pre-requisite to the `run_server` target so that Docker containers can be run with one command.
Added `set -e` to `createdb.sh` for better debugging of failed GoCD builds.