Removed use of venv for S3 operations.
Removed partly because it was not working on Windows 7 for Nayi and partly because it seemed unnecessary. The login step assumes that the AWS-CLI is already installed as it does not use a venv and that target is a pre-requisite for the `get_sql_dump` target affected by this commit.