Fixed `run_server` rule failing locally.
For some reason the SQL file wasn't being added to the database container. I believe this was caused by the build step not containing the override compose file that adds the SQL file and then the subsequent `up` commands using the faulty built container rather than detecting that they need to re-build.

Also changed the method of adding the SQL file to a volume mount rather than baking it into a new image. This is because the newly created image with the SQL file baked in was actually overwriting the local PostgreSQL image due to the way that the `build` and `image` properties interact with one another (they both exist due to the override compose files).